% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_crf_data_upload.R
\name{crfDataUploadUI}
\alias{crfDataUploadUI}
\title{crfDataUploadUI}
\usage{
crfDataUploadUI(id)
}
\arguments{
\item{id}{The NS id argument required for the module}
}
\value{
A shiny module UI object
}
\description{
crfDataUploadUI
}
