% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{df_to_rowslist}
\alias{df_to_rowslist}
\title{df_to_rowslist}
\usage{
df_to_rowslist(df, ...)
}
\arguments{
\item{df}{A dataframe}

\item{...}{any additional arguments passed along.
(Used by \link\link{list_of_dfs_to_rowslist})}
}
\value{
List of row vectors
}
\description{
Converts a dataframe to a list of row vectors.
}
\examples{
df_to_rowslist(mtcars)
}
