% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{layout}
\alias{layout}
\title{layout}
\usage{
layout(mainUI, header, navigation, footer)
}
\arguments{
\item{mainUI}{The main UI object}

\item{header}{header content}

\item{navigation}{navigation content}

\item{footer}{footer content}
}
\value{
A UI obbject
}
\description{
layout
}
