% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_of_dfs_to_rowslist}
\alias{list_of_dfs_to_rowslist}
\title{list_of_dfs_to_rowslist}
\usage{
list_of_dfs_to_rowslist(list_dfs)
}
\arguments{
\item{list_dfs}{List of dataframes}
}
\value{
A grouped list of row vectors
}
\description{
Conversts a list of dataframes into a grouped list of dataframe row vectors.
}
\examples{
list(one = mtcars, two = mtcars) \%>\% list_of_dfs_to_rowslist()
}
