% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{makeCard}
\alias{makeCard}
\title{makeCard}
\usage{
makeCard(title, content, size = 12, style = "")
}
\arguments{
\item{title}{Title of the card}

\item{content}{The copntetnts of the card}

\item{size}{CSS size}

\item{style}{CSS styles}
}
\value{
A UI object
}
\description{
makeCard
}
