% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{makePage}
\alias{makePage}
\title{makePage}
\usage{
makePage(title, subtitle, contents)
}
\arguments{
\item{title}{Page title}

\item{subtitle}{Page subtitle}

\item{contents}{Contents of the page}
}
\value{
A UI object
}
\description{
makePage
}
