% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_query_manager.R
\name{queryManagerServer}
\alias{queryManagerServer}
\title{crfDataUploadServer}
\usage{
queryManagerServer(id, crf_data)
}
\arguments{
\item{id}{The NS id argument required for the module}
}
\value{
A shiny module Server object
}
\description{
crfDataUploadServer
}
