% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_query_manager.R
\name{queryManagerUI}
\alias{queryManagerUI}
\title{queryManagerUI}
\usage{
queryManagerUI(id)
}
\arguments{
\item{id}{The NS id argument required for the module}
}
\value{
A UI object
}
\description{
queryManagerUI
}
